completed building home page.
