+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             android:name=".AddTask"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
