--- app/src/main/java/com/example/taskmaster/MainActivity.java
@@ -1,33 +1,0 @@
-package com.example.taskmaster;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        // Set onclick button event handling to add a task
-        Button addTaskButton = MainActivity.this.findViewById(R.id.addTaskButton);
-
-        addTaskButton.setOnClickListener(view -> {
-            Intent goToAddTaskActivity = new Intent(MainActivity.this, AddTask.class);
-            startActivity(goToAddTaskActivity);
-        });
-
-        // Set onclick button event handling to all tasks
-        Button allTasksButton = MainActivity.this.findViewById(R.id.allTasksButton);
-
-        allTasksButton.setOnClickListener(view -> {
-            Intent goToAllTasksActivity = new Intent(MainActivity.this, AllTasks.class);
-            startActivity(goToAllTasksActivity);
-        });
-    }
-}
