+++ app/src/main/res/layout/recycler_view_item.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              xmlns:tools="http://schemas.android.com/tools"
+              android:orientation="horizontal"
+              android:padding="8dp"
+              android:layout_width="match_parent"
+              android:layout_height="wrap_content">
+  <ImageView
+      android:id="@+id/imageView1"
+      android:layout_width="50dp"
+      android:layout_height="50dp"
+      tools:src="@tools:sample/avatars"/>
+  <TextView
+      android:id="@+id/textView1"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_gravity="center"
+      android:padding="5dp"
+      tools:text="@tools:sample/lorem" />
+</LinearLayout>
